requirements.txt
# python 3.6
# cmake==3.20.2

gpy==1.9.8
tensorflow-gpu==1.14.0
tensorflow-probability==0.7.0
keras==2.2.5
face-recognition==1.3.0
matplotlib==3.1.1
opencv-python==4.1.2.30
numba==0.48.0
tensorflow-determinism==0.3.0
joblib==0.14.0
scipy==1.2.3
numpy==1.18.3